:lipstick: (Login) Move <Login/> to be under user/ dir
Closes #22
