+++ pages/index.tsx
@@ -1,12 +1,3 @@
-export default function Login() {
-  return (
-    <>
-      <h1 className="">login</h1>
-      <p>Email: </p>
-      <p>Password: </p>
-      <button role="button" name="Submit user">
-        login
-      </button>
-    </>
-  );
+export default function Home() {
+  return <h1>Homepage</h1>;
 }
