+++ pages/user/login.tsx
@@ -1,0 +1,10 @@
+export default function Login() {
+  return (
+    <>
+      <h1>login</h1>
+      <p>Email</p>
+      <p>Password</p>
+      <button role="button">login</button>
+    </>
+  );
+}
